Set options on Faraday response object correctly
When you're making an HTTP request using Faraday and wish to customize
the request object, you *can* in theory override all of the options on
that object, but it doesn't work so well in practice. Instead, you need
to set the options on that object individually.

In other words, instead of saying

```
request.options = { foo: "bar" }
```

you say

```
request.options.foo = "bar"
```